Refactor project structure by moving database-related files into a dedicated 'database' directory and updating include paths accordingly
