+++ src/database/dbanalyzer.cpp
