+++ src/database/dbanalyzer.h
