+++ src/mainwindow.h
@@ -6,11 +6,11 @@
 #include <qfuturewatcher.h>
 
 #include "threading/cancellation.h"
-#include "dbanalyzer.h"
-#include "dbexport.h"
-#include "dbquery.h"
+#include "database/dbanalyzer.h"
+#include "database/dbexport.h"
+#include "database/dbquery.h"
 #include "highlighter.h"
-#include "dbtree.h"
+#include "database/dbtree.h"
 
 namespace Ui {
     class MainWindow;
